Refactor weight and input quantizers in separate files
